interfaces: move naming functions to dedicated file
This patch moves functions related to naming things to a dedicated file
in anticipation of unification and simplification of the agreement with
the security team.
