+++ interfaces/security.go
@@ -25,28 +25,6 @@ import (
 	"strings"
 )
 
-// WrapperNameForApp returns the name of the wrapper for a given application.
-//
-// A wrapper is a generated helper executable that assists in setting up
-// environment for running a particular application.
-//
-// In general, the wrapper has the form: "$snap.$app". When both snap name and
-// app name are the same then the tag is simplified to just "$snap".
-func WrapperNameForApp(snapName, appName string) string {
-	if appName == snapName {
-		return snapName
-	}
-	return fmt.Sprintf("%s.%s", snapName, appName)
-}
-
-// SecurityTagForApp returns the unified tag used for all security systems.
-//
-// In general, the tag has the form: "$snap.$app.snap". When both snap name and
-// app name are the same then the tag is simplified to just "$snap.snap".
-func SecurityTagForApp(snapName, appName string) string {
-	return fmt.Sprintf("%s.snap", WrapperNameForApp(snapName, appName))
-}
-
 // securityHelper is an interface for common aspects of generating security files.
 type securityHelper interface {
 	securitySystem() SecuritySystem
