+++ interfaces/security_test.go
@@ -78,20 +78,6 @@ func (s *SecuritySuite) prepareFixtureWithInterface(c *C, i Interface) {
 	c.Assert(err, IsNil)
 }
 
-// Tests for WrapperNameForApp()
-
-func (s *SecuritySuite) TestWrapperNameForApp(c *C) {
-	c.Assert(WrapperNameForApp("snap", "app"), Equals, "snap.app")
-	c.Assert(WrapperNameForApp("foo", "foo"), Equals, "foo")
-}
-
-// Tests for SecurityTagForApp()
-
-func (s *SecuritySuite) TestSecurityTagForApp(c *C) {
-	c.Assert(SecurityTagForApp("snap", "app"), Equals, "snap.app.snap")
-	c.Assert(SecurityTagForApp("foo", "foo"), Equals, "foo.snap")
-}
-
 // Tests for appArmor
 
 func (s *SecuritySuite) TestAppArmorPlugPermissions(c *C) {
